# doc/openapi.yaml
openapi: 3.0.0
info:
  title: API pour la création d'offres de stage sur https://stagedeseconde.1jeune1solution.gouv.fr/
  description: API pour la création d'offres de stages à destination des élèves de
    seconde générale et technologique
  contact:
    email: etienne.weil@monstagedetroisieme.fr
  version: 1.0.0
servers:
- url: http://localhost:3000/api
  description: SwaggerHub API Auto Mocking
security:
- bearerAuth: []
tags:
- name: developers
  description: Operations accessibles aux développeurs disposant d'une clé API. Celle-ci
    peut être fournie sur demande à l'adresse email contact @ stagedeseconde.education.gouv.fr
paths:
  /internship_offers:
    get:
      tags:
      - developers
      summary: liste (index) les offres déposées
      description: |
        Les offres retournées représentent l'ensemble des offres du site https://stagedeseconde.1jeune1solution.gouv.fr/ .
      operationId: internship_offers_index
      responses:
        "200":
          description: Offre de stage créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipOffers'
        "401":
          description: l'authentification a échoué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_401'
        "404":
          description: Un argument de la requête fait échouer la requête.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404_1'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Developers
    post:
      tags:
      - developers
      summary: ajoute une offre de stage d'un partenaire aux offres proposées en propre
        par les équipes 1élève1stage
      description: |
        Les offres de stages décrits ci-dessous décrivent les offres réservées aux classes de **seconde générale et technologique**
      operationId: internship_offers
      requestBody:
        $ref: '#/components/requestBodies/InternshipOffer'
      responses:
        "201":
          description: Offre de stage créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipOffer'
        "401":
          description: l'authentification a échoué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_401'
        "403":
          description: L'utilisateur n'est pas autorisé à supprimer cette offre de
            stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_403'
        "404":
          description: Un argument de la requête fait échouer la requête.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
        "406":
          description: Un ou plusieurs paramètres de requête sont invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_406'
        "409":
          description: Conflit. Une offre de stage avec le même remote_id existe déjà
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_409'
        "422":
          description: Un argument au moins de la requête fait échouer la requête
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_422'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Developers
  /internship_offers/{remote_id}:
    delete:
      tags:
      - developers
      summary: Supprime une offre de stage d'un partenaire des offres proposées sur
        le site https://stagedeseconde.1jeune1solution.gouv.fr/
      description: |
        Suppression des offres
      operationId: internship_offers_delete
      parameters:
      - name: remote_id
        in: path
        description: id de l'offre de stage
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: remote-123
      responses:
        "200":
          description: Offre de stage supprimée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipOffer'
        "401":
          description: l'authentification a échoué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_401'
        "403":
          description: L'utilisateur n'est pas autorisé à supprimer cette offre de
            stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_403'
        "404":
          description: Un argument de la requête fait échouer la requête.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
        "422":
          description: Un argument de la requête fait échouer la requête.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_422'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Developers
    patch:
      tags:
      - developers
      summary: met à jour une offre de stage d'un partenaire aux offres proposées
        en propre par les équipes 1élève1stage
      description: |
        Mise à jour des offres . Possibilité de désactiver son offre en passant le paramètre published_at à null
      operationId: internship_offers_updatte
      parameters:
      - name: remote_id
        in: path
        description: "clé étrangère unique , id de l'offre de stage chez le partenaire"
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: remote-123
      requestBody:
        $ref: '#/components/requestBodies/InternshipOfferPatch'
      responses:
        "200":
          description: Offre de stage mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipOffer'
        "401":
          description: l'authentification a échoué
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_401'
        "403":
          description: L'utilisateur n'est pas autorisé à supprimer cette offre de
            stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_403'
        "404":
          description: Un argument de la requête fait échouer la requête.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
        "406":
          description: Un ou plusieurs paramètres de requête sont invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_406'
        "422":
          description: Un argument au moins de la requête fait échouer la requête
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_422'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Developers
  /internship_offers/search:
    get:
      tags:
        - developers
      summary: Recherche des offres de stage
      description: Recherche des offres de stage déposées par le partenaire.

      parameters:
        - in: query
          name: latitude
          schema:
            type: number
          required: false
          example: 48.8566
        - in: query
          name: longitude
          schema:
            type: number
          required: false
          example: 2.3522
        - in: query
          name: radius
          schema:
            type: integer
          required: false
          example: 40000
        - in: query
          name: keyword
          schema:
            type: string
          required: false
          example: "Game designer"

      responses:
        '200':
          description: Offres de stages du partenaire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipOffers'
      security:
      - bearerAuth: []
      x-swagger-router-controller: Developers
  /sectors:
    get:
      tags:
      - developers
      summary: Récupère la liste des secteurs économiques
      description: "Secteur économique. Voir la liste complète ici : [https://github.com/betagouv/monstagedeseconde/blob/staging/doc/README.md#secteurs-dactivité\
        ](https://github.com/betagouv/monstagedeseconde/blob/staging/doc/README.md#secteurs-dactivité\
        )"
      operationId: sectors
      responses:
        "200":
          description: Liste des secteurs économiques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sector'
                x-content-type: application/json
      security:
      - bearerAuth: []
      x-swagger-router-controller: Developers
components:
  schemas:
    InternshipOffers:
      type: array
      items:
        $ref: '#/components/schemas/InternshipOffer'
    InternshipOffer:
      required:
      - city
      - coordinates
      - employer_name
      - is_public
      - max_candidates
      - period
      - remote_id
      - sector_uuid
      - street
      - title
      type: object
      properties:
        remote_id:
          type: string
          description: Identifiant de l'offre de stage chez le partenaire
          example: A156548-H
        description:
          maxLength: 500
          type: string
          description: Description de l'offre de stage
        title:
          type: string
          example: Stage d'observation du métier de chef de service
        employer_name:
          type: string
          example: BNP Paribas
        employer_description:
          maxLength: 275
          type: string
          example: "Créateur de lotions, de parfums et de produits cosmétiques, embaumeur"
        street:
          type: string
          example: 16 rue de la paix
        city:
          type: string
          example: Paris
        zipcode:
          type: string
          example: "75001"
        employer_website:
          type: string
          example: http://www.acnee-corporation.fr
        lunch_break:
          type: string
          description: Horaires de pause déjeuner
          example: 12h-14h
        daily_hours:
          $ref: '#/components/schemas/InternshipOffer_daily_hours'
        coordinates:
          $ref: '#/components/schemas/InternshipOffer_coordinates'
        period:
          type: integer
          description: Semaines de stage proposées où 0 => 2 semaines - du 17 au 28
            juin 2024; 1 => 1 semaine - du 17 au 21 juin 2024; 2 => 1 semaine - du
            24 au 28 juin 2024
          example: 1
          enum:
          - 0
          - 1
          - 2
        permalink:
          type: string
          description: Site de l'employeur
          example: http://www.stagechezemployeur.fr
        max_candidates:
          type: integer
          description: Nombre maximum de candidats pouvant postuler à cette offre
            de stage sur l'ensemble des semaines proposées
          example: 1
        is_public:
          type: boolean
          description: "true si l'offre de stage est issue d'une administration publique,\
            \ false si elle est issue d'une entreprise privée"
          example: true
        sector_uuid:
          $ref: '#/components/schemas/Sector'
      description: offre de stage
      example:
        remote_id: A123-12
        period: 0
        city: Paris
        coordinates:
          latitude: 48.866667
          longitude: 2.333333
        description: Stage sur le thème de la logistique et de la supply chain
        employer_description: "Créateur de lotions, de parfums et de produits cosmé\
          tiques, embaumeur"
        employer_name: BNP Paribas
        title: Stage d'observation du métier de chef de service
        lunch_break: 12h-14h
        zipcode: "75001"
        max_candidates: 1
        employer_website: http://www.acnee-corporation.fr
        sector_uuid:
          sector_uuid: s20
          name: Mode
          id: 1
        street: 16 rue de la paix
        is_public: true
        permalink: http://www.stagechezemployeur.fr
        daily_hours:
          type: object
          description: Horaires de stage
          properties:
            lundi:
              type: array
              items:
                type: string
              example:
              - "9:00"
              - "18:00"
            mardi:
              type: array
              items:
                type: string
              example:
              - "9:00"
              - "18:00"
            mercredi:
              type: array
              items:
                type: string
              example:
              - "9:00"
              - "18:00"
            jeudi:
              type: array
              items:
                type: string
              example:
              - "9:00"
              - "18:00"
            vendredi:
              type: array
              items:
                type: string
              example:
              - "9:00"
              - "18:00"
    InternshipOfferPatch:
      allOf:
      - type: object
        properties:
          published_at:
            type: string
            format: date-time
      - $ref: '#/components/schemas/InternshipOffer'
    Sector:
      type: object
      properties:
        id:
          type: number
          example: 1
        sector_uuid:
          type: string
          format: uuid
        name:
          type: string
          example: Mode
      description: Secteur économique.
      example:
        sector_uuid: s20
        name: Mode
        id: 1
    inline_response_401:
      type: object
      properties:
        code:
          type: string
          example: UNAUTHORIZED
        error:
          type: string
          example: access denied
        status:
          type: string
          example: unauthorized
    inline_response_403:
      type: object
      properties:
        code:
          type: string
          example: FORBIDDEN
        error:
          type: string
          example: access denied
        status:
          type: string
          example: forbidden
    inline_response_404:
      type: object
      properties:
        code:
          type: string
          example: NOT_FOUND
        error:
          type: string
          example: can't find internship_offer with this remote_id
        status:
          type: string
          example: not_found
    inline_response_406:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        error:
          type: string
          example: Erreur de validation sur l'objet
        status:
          type: string
          example: conflict
    inline_response_409:
      type: object
      properties:
        code:
          type: string
          example: DUPLICATE_WEEKY_FRAMED
        error:
          type: string
          example: period with this remote_id ("A123-2") already exists.
        status:
          type: string
          example: conflict
    inline_response_422:
      type: object
      properties:
        code:
          type: string
          example: BAD_ARGUMENT
        error:
          type: string
          example: Erreur sur la requête due à un un paramètre invalide
        status:
          type: string
          example: unprocessable_entity
    inline_response_404_1:
      type: object
      properties:
        code:
          type: string
          example: NOT_FOUND
        error:
          type: string
          example: can't find internship_offer with this remote_id
        status:
          type: string
          example: not_found
    inline_response_429:
      type: object
      properties:
        code:
          type: string
          example: TOO_MANY_REQUESTS
        error:
          type: string
          example: request limit exceeded (100 per minutes)
        status:
          type: string
          example: too_many_requests
    InternshipOffer_coordinates:
      type: object
      properties:
        latitude:
          type: number
          example: 48.866667
        longitude:
          type: number
          example: 2.333333
      description: Coordonnées géographiques du lieu de stage
    InternshipOffer_daily_hours:
      type: object
      properties:
        lundi:
          type: array
          example:
          - "9:00"
          - "18:00"
          items:
            type: string
        mardi:
          type: array
          example:
          - "9:00"
          - "18:00"
          items:
            type: string
        mercredi:
          type: array
          example:
          - "9:00"
          - "18:00"
          items:
            type: string
        jeudi:
          type: array
          example:
          - "9:00"
          - "18:00"
          items:
            type: string
        vendredi:
          type: array
          example:
          - "9:00"
          - "18:00"
          items:
            type: string
      description: Horaires de stage
  responses:
    Unauthorized:
      description: l'authentification a échoué
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_response_401'
  requestBodies:
    InternshipOffer:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternshipOffer'
    InternshipOfferPatch:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternshipOfferPatch'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT